+++ content/!del/geometry/ClosestPair.h
@@ -11,15 +11,15 @@
 
 #include "Point.h"
 
-typedef Point<ll> P;
+typedef Point<int> P;
 pair<P, P> closest(vector<P> v) {
 	assert(sz(v) > 1);
 	set<P> S;
 	sort(all(v), [](P a, P b) { return a.y < b.y; });
-	pair<ll, pair<P, P>> ret{LLONG_MAX, {P(), P()}};
+	pair<int, pair<P, P>> ret{LLONG_MAX, {P(), P()}};
 	int j = 0;
 	for (P p : v) {
-		P d{1 + (ll)sqrt(ret.first), 0};
+		P d{1 + (int)sqrt(ret.first), 0};
 		while (v[j].y <= p.y - d.x) S.erase(v[j++]);
 		auto lo = S.lower_bound(p - d), hi = S.upper_bound(p + d);
 		for (; lo != hi; ++lo)
