+++ content/combinatorial/multinomial.h
@@ -7,8 +7,8 @@
  */
 #pragma once
 
-ll multinomial(vi& v) {
-	ll c = 1, m = v.empty() ? 1 : v[0];
+int multinomial(vi& v) {
+	int c = 1, m = v.empty() ? 1 : v[0];
 	rep(i,1,sz(v)) rep(j,0,v[i]) c = c * ++m / (j+1);
 	return c;
 }
