+++ content/data-structures/LineContainer.h
@@ -11,15 +11,15 @@
 #pragma once
 
 struct Line {
-	mutable ll k, m, p;
+	mutable int k, m, p;
 	bool operator<(const Line& o) const { return k < o.k; }
-	bool operator<(ll x) const { return p < x; }
+	bool operator<(int x) const { return p < x; }
 };
 
 struct LineContainer : multiset<Line, less<>> {
 	// (for doubles, use inf = 1/.0, div(a,b) = a/b)
-	static const ll inf = LLONG_MAX;
-	ll div(ll a, ll b) { // floored division
+	static const int inf = LLONG_MAX;
+	int div(int a, int b) { // floored division
 		return a / b - ((a ^ b) < 0 && a % b); }
 	bool isect(iterator x, iterator y) {
 		if (y == end()) return x->p = inf, 0;
@@ -27,14 +27,14 @@ struct LineContainer : multiset<Line, less<>> {
 		else x->p = div(y->m - x->m, x->k - y->k);
 		return x->p >= y->p;
 	}
-	void add(ll k, ll m) {
+	void add(int k, int m) {
 		auto z = insert({k, m, 0}), y = z++, x = y;
 		while (isect(y, z)) z = erase(z);
 		if (x != begin() && isect(--x, y)) isect(x, y = erase(y));
 		while ((y = x) != begin() && (--x)->p >= y->p)
 			isect(x, erase(y));
 	}
-	ll query(ll x) {
+	int query(int x) {
 		assert(!empty());
 		auto l = *lower_bound(x);
 		return l.k * x + l.m;
