+++ content/graph/del/DirectedMST.h
@@ -13,11 +13,11 @@
 
 #include "../data-structures/UnionFindRollback.h"
 
-struct Edge { int a, b; ll w; };
+struct Edge { int a, b; int w; };
 struct Node { /// lazy skew heap node
 	Edge key;
 	Node *l, *r;
-	ll delta;
+	int delta;
 	void prop() {
 		key.w += delta;
 		if (l) l->delta += delta;
@@ -35,11 +35,11 @@ Node *merge(Node *a, Node *b) {
 }
 void pop(Node*& a) { a->prop(); a = merge(a->l, a->r); }
 
-pair<ll, vi> dmst(int n, int r, vector<Edge>& g) {
+pair<int, vi> dmst(int n, int r, vector<Edge>& g) {
 	RollbackUF uf(n);
 	vector<Node*> heap(n);
 	for (Edge e : g) heap[e.b] = merge(heap[e.b], new Node{e});
-	ll res = 0;
+	int res = 0;
 	vi seen(n, -1), path(n), par(n);
 	seen[r] = r;
 	vector<Edge> Q(n), in(n, {-1,-1}), comp;
