+++ content/graph/del/GeneralMatching.h
@@ -13,7 +13,7 @@
 #include "../numerical/MatrixInverse-mod.h"
 
 vector<pii> generalMatching(int N, vector<pii>& ed) {
-	vector<vector<ll>> mat(N, vector<ll>(N)), A;
+	vector<vector<int>> mat(N, vector<int>(N)), A;
 	for (pii pa : ed) {
 		int a = pa.first, b = pa.second, r = rand() % mod;
 		mat[a][b] = r, mat[b][a] = (mod - r) % mod;
@@ -23,7 +23,7 @@ vector<pii> generalMatching(int N, vector<pii>& ed) {
 	assert(r % 2 == 0);
 
 	if (M != N) do {
-		mat.resize(M, vector<ll>(M));
+		mat.resize(M, vector<int>(M));
 		rep(i,0,N) {
 			mat[i].resize(M);
 			rep(j,N,M) {
@@ -42,9 +42,9 @@ vector<pii> generalMatching(int N, vector<pii>& ed) {
 		if (fj < N) ret.emplace_back(fi, fj);
 		has[fi] = has[fj] = 0;
 		rep(sw,0,2) {
-			ll a = modpow(A[fi][fj], mod-2);
+			int a = modpow(A[fi][fj], mod-2);
 			rep(i,0,M) if (has[i] && A[i][fj]) {
-				ll b = A[i][fj] * a % mod;
+				int b = A[i][fj] * a % mod;
 				rep(j,0,M) A[i][j] = (A[i][j] - A[fi][j] * b) % mod;
 			}
 			swap(fi,fj);
