+++ content/graph/del/MinCostMaxFlow.h
@@ -12,22 +12,22 @@
 
 // #include <bits/extc++.h> /// include-line, keep-include
 
-const ll INF = numeric_limits<ll>::max() / 4;
+const int INF = numeric_limits<int>::max() / 4;
 
 struct MCMF {
 	struct edge {
 		int from, to, rev;
-		ll cap, cost, flow;
+		int cap, cost, flow;
 	};
 	int N;
 	vector<vector<edge>> ed;
 	vi seen;
-	vector<ll> dist, pi;
+	vector<int> dist, pi;
 	vector<edge*> par;
 
 	MCMF(int N) : N(N), ed(N), seen(N), dist(N), pi(N), par(N) {}
 
-	void addEdge(int from, int to, ll cap, ll cost) {
+	void addEdge(int from, int to, int cap, int cost) {
 		if (from == to) return;
 		ed[from].push_back(edge{ from,to,sz(ed[to]),cap,cost,0 });
 		ed[to].push_back(edge{ to,from,sz(ed[from])-1,0,-cost,0 });
@@ -36,9 +36,9 @@ struct MCMF {
 	void path(int s) {
 		fill(all(seen), 0);
 		fill(all(dist), INF);
-		dist[s] = 0; ll di;
+		dist[s] = 0; int di;
 
-		__gnu_pbds::priority_queue<pair<ll, int>> q;
+		__gnu_pbds::priority_queue<pair<int, int>> q;
 		vector<decltype(q)::point_iterator> its(N);
 		q.push({ 0, s });
 
@@ -46,7 +46,7 @@ struct MCMF {
 			s = q.top().second; q.pop();
 			seen[s] = 1; di = dist[s] + pi[s];
 			for (edge& e : ed[s]) if (!seen[e.to]) {
-				ll val = di - pi[e.to] + e.cost;
+				int val = di - pi[e.to] + e.cost;
 				if (e.cap - e.flow > 0 && val < dist[e.to]) {
 					dist[e.to] = val;
 					par[e.to] = &e;
@@ -60,10 +60,10 @@ struct MCMF {
 		rep(i,0,N) pi[i] = min(pi[i] + dist[i], INF);
 	}
 
-	pair<ll, ll> maxflow(int s, int t) {
-		ll totflow = 0, totcost = 0;
+	pair<int, int> maxflow(int s, int t) {
+		int totflow = 0, totcost = 0;
 		while (path(s), seen[t]) {
-			ll fl = INF;
+			int fl = INF;
 			for (edge* x = par[t]; x; x = par[x->from])
 				fl = min(fl, x->cap - x->flow);
 
@@ -80,7 +80,7 @@ struct MCMF {
 	// If some costs can be negative, call this before maxflow:
 	void setpi(int s) { // (otherwise, leave this out)
 		fill(all(pi), INF); pi[s] = 0;
-		int it = N, ch = 1; ll v;
+		int it = N, ch = 1; int v;
 		while (ch-- && it--)
 			rep(i,0,N) if (pi[i] != INF)
 			  for (edge& e : ed[i]) if (e.cap)
