+++ content/number-theory/ModMulLL.h
@@ -16,13 +16,13 @@
  */
 #pragma once
 
-typedef unsigned long long ull;
-ull modmul(ull a, ull b, ull M) {
-	ll ret = a * b - M * ull(1.L / M * a * b);
-	return ret + M * (ret < 0) - M * (ret >= (ll)M);
+typedef unsigned long long uint;
+uint modmul(uint a, uint b, uint M) {
+	int ret = a * b - M * uint(1.L / M * a * b);
+	return ret + M * (ret < 0) - M * (ret >= (int)M);
 }
-ull modpow(ull b, ull e, ull mod) {
-	ull ans = 1;
+uint modpow(uint b, uint e, uint mod) {
+	uint ans = 1;
 	for (; e; b = modmul(b, b, mod), e /= 2)
 		if (e & 1) ans = modmul(ans, b, mod);
 	return ans;
