+++ content/number-theory/ModSum.h
@@ -12,19 +12,19 @@
  */
 #pragma once
 
-typedef unsigned long long ull;
-ull sumsq(ull to) { return to / 2 * ((to-1) | 1); }
+typedef unsigned long long uint;
+uint sumsq(uint to) { return to / 2 * ((to-1) | 1); }
 /// ^ written in a weird way to deal with overflows correctly
 
-ull divsum(ull to, ull c, ull k, ull m) {
-	ull res = k / m * sumsq(to) + c / m * to;
+uint divsum(uint to, uint c, uint k, uint m) {
+	uint res = k / m * sumsq(to) + c / m * to;
 	k %= m; c %= m;
 	if (!k) return res;
-	ull to2 = (to * k + c) / m;
+	uint to2 = (to * k + c) / m;
 	return res + (to - 1) * to2 - divsum(to2, m-1 - c, m, k);
 }
 
-ll modsum(ull to, ll c, ll k, ll m) {
+int modsum(uint to, int c, int k, int m) {
 	c = ((c % m) + m) % m;
 	k = ((k % m) + m) % m;
 	return to * c + k * sumsq(to) - m * divsum(to, c, k, m);
