+++ content/number-theory/del/FracBinarySearch.h
@@ -11,16 +11,16 @@
  * Status: stress-tested for n <= 300
  */
 
-struct Frac { ll p, q; };
+struct Frac { int p, q; };
 
 template<class F>
-Frac fracBS(F f, ll N) {
+Frac fracBS(F f, int N) {
 	bool dir = 1, A = 1, B = 1;
 	Frac lo{0, 1}, hi{1, 1}; // Set hi to 1/0 to search (0, N]
 	if (f(lo)) return lo;
 	assert(f(hi));
 	while (A || B) {
-		ll adv = 0, step = 1; // move hi if dir, else lo
+		int adv = 0, step = 1; // move hi if dir, else lo
 		for (int si = 0; step; (step *= 2) >>= si) {
 			adv += step;
 			Frac mid{lo.p * adv + hi.p, lo.q * adv + hi.q};
