+++ content/various/del/SIMD.h
@@ -41,8 +41,8 @@ mi one() { return _mm256_set1_epi32(-1); }
 bool all_zero(mi m) { return _mm256_testz_si256(m, m); }
 bool all_one(mi m) { return _mm256_testc_si256(m, one()); }
 
-ll example_filteredDotProduct(int n, short* a, short* b) {
-	int i = 0; ll r = 0;
+int example_filteredDotProduct(int n, short* a, short* b) {
+	int i = 0; int r = 0;
 	mi zero = _mm256_setzero_si256(), acc = zero;
 	while (i + 16 <= n) {
 		mi va = L(a[i]), vb = L(b[i]); i += 16;
@@ -51,7 +51,7 @@ ll example_filteredDotProduct(int n, short* a, short* b) {
 		acc = _mm256_add_epi64(_mm256_unpacklo_epi32(vp, zero),
 			_mm256_add_epi64(acc, _mm256_unpackhi_epi32(vp, zero)));
 	}
-	union {ll v[4]; mi m;} u; u.m = acc; rep(i,0,4) r += u.v[i];
+	union {int v[4]; mi m;} u; u.m = acc; rep(i,0,4) r += u.v[i];
 	for (;i<n;++i) if (a[i] < b[i]) r += a[i]*b[i]; // <- equiv
 	return r;
 }
